# src/main/java/com/hemit/controllers/OfferController.java
package com.hemit.controllers;

import com.hemit.models.Offer;
import com.hemit.repositories.OfferRepository;
import io.swagger.annotations.ApiOperation;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.vertx.core.cli.annotations.Summary;
import org.bson.types.ObjectId;
import org.eclipse.microprofile.openapi.annotations.responses.APIResponseSchema;
import org.eclipse.microprofile.openapi.annotations.tags.Tag;

import javax.inject.Inject;
import javax.ws.rs.*;
import javax.ws.rs.core.Response;
import java.util.List;

@Tag(name="offer", description="Gestion des offer")

@Path("/offers")
public class OfferController {

    private final OfferRepository offerRepository;

    @Inject
    public OfferController(OfferRepository offerRepository) {
        this.offerRepository = offerRepository;
    }


    @POST
    public Response create(Offer offer) {
        offerRepository.persist(offer);
        return Response.status(Response.Status.CREATED).build();
    }

    @PUT
    @Path("/{id}")
    public void update(@PathParam("id") String id, Offer offer) {
        offer.id = new ObjectId(id);
        offerRepository.update(offer);
    }

    @DELETE
    @Path("/{id}")
    public void delete(@PathParam("id") String id) {
        Offer offer = offerRepository.findById(new ObjectId(id));
        offerRepository.delete(offer);
    }

    @GET
    @Path("/{id}")
    public Offer get(@PathParam("id") String id) {
        return offerRepository.findById(new ObjectId(id));
    }

    @GET
    public List<Offer> list() {
        return offerRepository.listAll();
    }

    @GET
    @Path("/count")
    public Long count() {
        return offerRepository.count();
    }

}